refactor(proxy): extract affinity routing helpers
Move prompt-cache and sticky-session affinity helpers out of the ProxyService monolith into a focused proxy.affinity module. Update API turn-state helpers and tests to use the new module seam without changing routing behavior.
